requirements.txt
certifi==2020.6.20
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.0.12
django-import-export==2.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
numpy==1.14.5
odfpy==1.4.1
openpyxl==3.0.5
pandas==0.23.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
scipy==1.1.0
six==1.15.0
tablib==2.0.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
